clean up header mod
