+++ actix-http/src/header/common/if_range.rs
@@ -110,7 +110,8 @@ mod test_if_range {
     use super::IfRange as HeaderField;
     use crate::header::*;
     use std::str;
+
     test_header!(test1, vec![b"Sat, 29 Oct 1994 19:43:31 GMT"]);
-    test_header!(test2, vec![b"\"xyzzy\""]);
+    test_header!(test2, vec![b"\"abc\""]);
     test_header!(test3, vec![b"this-is-invalid"], None::<IfRange>);
 }
