+++ actix-http/src/header/shared/mod.rs
@@ -1,14 +1,16 @@
 //! Originally taken from `hyper::header::shared`.
 
-pub use self::charset::Charset;
-pub use self::encoding::Encoding;
-pub use self::entity::EntityTag;
-pub use self::httpdate::HttpDate;
-pub use self::quality_item::{q, qitem, Quality, QualityItem};
-pub use language_tags::LanguageTag;
-
 mod charset;
 mod encoding;
 mod entity;
+mod extended;
 mod httpdate;
 mod quality_item;
+
+pub use self::charset::Charset;
+pub use self::encoding::Encoding;
+pub use self::entity::EntityTag;
+pub use self::extended::{parse_extended_value, ExtendedValue};
+pub use self::httpdate::HttpDate;
+pub use self::quality_item::{q, qitem, Quality, QualityItem};
+pub use language_tags::LanguageTag;
